Update MS_Office_2016_PKGBuilder.sh
Change behavior when no exclude parameters are set.

+++ Office2016PackageBuilder/MS_Office_2016_PKGBuilder.sh
@@ -186,7 +186,13 @@ else
 fi
 
 # Install Microsoft Office 2016
-/usr/sbin/installer -dumplog -verbose -pkg "$WORKING_DIR/Microsoft_Office_2016_Installer.pkg" -applyChoiceChangesXML "$WORKING_DIR/choices.xml" -target "$3"   2>&1 | tee -a $LOGFILE
+if [[ -e "$WORKING_DIR/choices.xml" ]]; then
+    echo "InstallerChoices file found! Using this it!"      2>&1 | tee -a $LOGFILE
+    /usr/sbin/installer -dumplog -verbose -pkg "$WORKING_DIR/Microsoft_Office_2016_Installer.pkg" -applyChoiceChangesXML "$WORKING_DIR/choices.xml" -target "$3"   2>&1 | tee -a $LOGFILE
+else
+    echo "InstallerChoices file not found! Continuing!"      2>&1 | tee -a $LOGFILE
+    /usr/sbin/installer -dumplog -verbose -pkg "$WORKING_DIR/Microsoft_Office_2016_Installer.pkg" -target "$3"   2>&1 | tee -a $LOGFILE
+fi
 
 # Install the Microsoft Office 2016 Volume License file from
 /usr/sbin/installer -dumplog -verbose -pkg "$WORKING_DIR/Microsoft_Office_2016_VL_Serializer.pkg" -target "$3"   2>&1 | tee -a $LOGFILE
